One final rearrangement to addData function
